refactor: 调整工具函数位置，移除 OptionalPathTask
