+++ src/nonebot_plugin_parser/parsers/task.py
@@ -47,44 +47,3 @@ class PathTask:
             return f"PathTask(path={self._path.name})"
         else:
             return f"PathTask(task={self._task.get_name()}, done={self._task.done()})"
-
-
-class OptionalPathTask:
-    __slots__ = ("_path_task",)
-
-    def __init__(
-        self,
-        path_task: Task[Path] | PathTask | Coroutine[Any, Any, Path] | None = None,
-    ):
-        self._path_task: PathTask | None = None
-
-        if path_task is None:
-            pass
-        elif isinstance(path_task, PathTask):
-            self._path_task = path_task
-        else:
-            self._path_task = PathTask(path_task)
-
-    async def get(self) -> Path | None:
-        if self._path_task is None:
-            return None
-        return await self._path_task.get()
-
-    async def safe_get(
-        self,
-        on_error: Callable[[Exception], None] | None = None,
-    ) -> Path | None:
-        """任务失败, 返回 None"""
-        if self._path_task is None:
-            return None
-        return await self._path_task.safe_get(on_error)
-
-    @property
-    async def uri(self) -> str | None:
-        path = await self.safe_get()
-        if path is not None:
-            return path.as_uri()
-        return None
-
-    def __repr__(self) -> str:
-        return f"{self._path_task}"
