+++ src/nonebot_plugin_parser/parsers/utils.py
@@ -1,0 +1,9 @@
+def fmt_duration(duration: float) -> str:
+    """格式化媒体时长，超过 1 小时后显示为 h:mm:ss。"""
+    total_seconds = max(int(duration), 0)
+    hours, remainder = divmod(total_seconds, 3600)
+    minutes, seconds = divmod(remainder, 60)
+
+    if hours:
+        return f"{hours}:{minutes:02d}:{seconds:02d}"
+    return f"{minutes}:{seconds:02d}"
