+++ src/nonebot_plugin_parser/utils.py
@@ -171,17 +171,6 @@ def fmt_size(file_path: Path) -> str:
     return f"大小: {file_path.stat().st_size / 1024 / 1024:.2f} MB"
 
 
-def fmt_duration(duration: float) -> str:
-    """格式化媒体时长，超过 1 小时后显示为 h:mm:ss。"""
-    total_seconds = max(int(duration), 0)
-    hours, remainder = divmod(total_seconds, 3600)
-    minutes, seconds = divmod(remainder, 60)
-
-    if hours:
-        return f"{hours}:{minutes:02d}:{seconds:02d}"
-    return f"{minutes}:{seconds:02d}"
-
-
 def generate_file_name(url: str, default_suffix: str = "") -> str:
     """根据 url 生成文件名"""
